Added Tool Output Info "Agent is working on..."
Added print to exec_stored_func tool so we know what function the LLM agent is trying to run.  "Agent is attempting to run cap_outliers_and_impute_nulls on the 'col3' column"

Added func_tester.py
This is a simple file that allows you to run/test a function from the static_lib on a given column or set of columns without doing an LLM agent calls. Now when creating new functions, we can run basic debugging in a python only space.